+++ src/render/descriptor_pool.cc
@@ -16,7 +16,7 @@ render::DescriptorPool::DescriptorPool(const DeviceConfiguration& device_cfg, ui
 
 	VkDescriptorPoolCreateInfo pool_info{};
 	pool_info.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-	pool_info.poolSizeCount = pool_sizes.size();
+	pool_info.poolSizeCount = u32(pool_sizes.size());
 	pool_info.pPoolSizes = pool_sizes.data();
 	pool_info.maxSets = uniform_set_cnt + sampler_set_cnt;
